refactor(Race): 경주 객체 역할 분리/ Response를 Race가 만들도록 변경
